Extract MoveObjectByMouse code to separate class
