+++ source/MRCommonPlugins/Basic/MRMoveObjectByMouse.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "MRViewer/MRStatePlugin.h"
+#include "MRViewer/MRMoveObjectByMouseImpl.h"
 #include "MRMesh/MRPlane3.h"
 #include "MRMesh/MRAffineXf3.h"
 #include "imgui.h"
@@ -19,31 +20,11 @@ public:
     virtual bool blocking() const override { return false; };
 
 private:
-    virtual bool onMouseDown_( MouseButton button, int modifier ) override;
+    virtual bool onMouseDown_( MouseButton btn, int modifiers ) override;
     virtual bool onMouseMove_( int x, int y ) override;
     virtual bool onMouseUp_( MouseButton btn, int modifiers ) override;
 
-    void setVisualizeVectors_( std::vector<Vector3f> worldPoints );
-
-    std::shared_ptr<VisualObject> obj_;
-
-    Vector3f worldStartPoint_;
-    Vector3f worldBboxCenter_;
-    Vector3f bboxCenter_;
-    AffineXf3f objWorldXf_;
-    float viewportStartPointZ_;
-    Plane3f rotationPlane_;
-
-    std::vector<ImVec2> visualizeVectors_;
-    float angle_ = 0.f;
-    float shift_ = 0.f;
-
-    enum class TransformMode
-    {
-        Translation,
-        Rotation,
-        None
-    } transformMode_ = TransformMode::None;
+    MoveObjectByMouseImpl moveByMouse_;
 };
 
 }
